# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Progi CAP API",
        "description": "API documentation for Progi CAP application modules",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API server"
        }
    ],
    "paths": {
        "/api/pending-students": {
            "get": {
                "tags": [
                    "Pending Students"
                ],
                "summary": "Liste des étudiants en attente",
                "description": "Récupère la liste paginée des étudiants en attente avec possibilité de filtrage",
                "operationId": "getPendingStudents",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "documents_submitted",
                                "approved",
                                "rejected"
                            ]
                        }
                    },
                    {
                        "name": "entry_level_id",
                        "in": "query",
                        "description": "Filtrer par niveau d'entrée",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des étudiants récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PendingStudent"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pending Students"
                ],
                "summary": "Créer un étudiant en attente",
                "description": "Crée un nouvel étudiant en attente d'inscription",
                "operationId": "createPendingStudent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "first_name",
                                    "last_name",
                                    "phone",
                                    "entry_level_id",
                                    "entry_diploma_id"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "student@example.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "entry_level_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "entry_diploma_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Étudiant créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Étudiant en attente créé avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PendingStudent"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginationMeta": {
                "title": "Pagination Meta",
                "description": "Métadonnées de pagination",
                "properties": {
                    "total": {
                        "description": "Nombre total d'éléments",
                        "type": "integer"
                    },
                    "per_page": {
                        "description": "Nombre d'éléments par page",
                        "type": "integer"
                    },
                    "current_page": {
                        "description": "Page actuelle",
                        "type": "integer"
                    },
                    "last_page": {
                        "description": "Dernière page",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Premier élément de la page",
                        "type": "integer"
                    },
                    "to": {
                        "description": "Dernier élément de la page",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "EntryDiploma": {
                "title": "Entry Diploma",
                "description": "Modèle représentant un diplôme d'entrée",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom du diplôme",
                        "type": "string"
                    },
                    "uuid": {
                        "description": "UUID unique",
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EntryLevel": {
                "title": "Entry Level",
                "description": "Modèle représentant un niveau d'entrée",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom du niveau",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description du niveau",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PendingStudent": {
                "title": "Pending Student",
                "description": "Modèle représentant un étudiant en attente d'inscription",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "email": {
                        "description": "Adresse email",
                        "type": "string",
                        "format": "email"
                    },
                    "first_name": {
                        "description": "Prénom",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Nom de famille",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Numéro de téléphone",
                        "type": "string"
                    },
                    "entry_level_id": {
                        "description": "ID du niveau d'entrée",
                        "type": "integer"
                    },
                    "entry_diploma_id": {
                        "description": "ID du diplôme d'entrée",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Statut de l'inscription",
                        "type": "string",
                        "enum": [
                            "pending",
                            "documents_submitted",
                            "approved",
                            "rejected"
                        ]
                    },
                    "submitted_at": {
                        "description": "Date de soumission",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    },
                    "entry_level": {
                        "$ref": "#/components/schemas/EntryLevel"
                    },
                    "entry_diploma": {
                        "$ref": "#/components/schemas/EntryDiploma"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}